Extracted find many records query
