+++ packages/twenty-front/src/modules/object-record/hooks/useExecuteQuickActionOnOneRecord.ts
@@ -3,6 +3,7 @@ import { useApolloClient } from '@apollo/client';
 import { getOperationName } from '@apollo/client/utilities';
 
 import { useObjectMetadataItem } from '@/object-metadata/hooks/useObjectMetadataItem';
+import { useFindManyRecordsQuery } from '@/object-record/hooks/useFindManyRecordsQuery';
 import { capitalize } from '~/utils/string/capitalize';
 
 type useExecuteQuickActionOnOneRecordProps = {
@@ -12,11 +13,12 @@ type useExecuteQuickActionOnOneRecordProps = {
 export const useExecuteQuickActionOnOneRecord = <T>({
   objectNameSingular,
 }: useExecuteQuickActionOnOneRecordProps) => {
-  const {
-    objectMetadataItem,
-    executeQuickActionOnOneRecordMutation,
-    findManyRecordsQuery,
-  } = useObjectMetadataItem({
+  const { objectMetadataItem, executeQuickActionOnOneRecordMutation } =
+    useObjectMetadataItem({
+      objectNameSingular,
+    });
+
+  const { findManyRecordsQuery } = useFindManyRecordsQuery({
     objectNameSingular,
   });
