+++ packages/twenty-front/src/modules/object-record/query-keys/types/QueryFields.ts
@@ -1,0 +1,1 @@
+export type QueryFields = Record<string, any>;
