+++ packages/twenty-front/src/modules/object-record/query-keys/types/QueryKey.ts
@@ -1,9 +1,10 @@
+import { QueryFields } from '@/object-record/query-keys/types/QueryFields';
 import { ObjectRecordQueryVariables } from '@/object-record/types/ObjectRecordQueryVariables';
 
 export type QueryKey = {
   objectNameSingular: string;
   variables: ObjectRecordQueryVariables;
   depth?: number;
-  fields?: Record<string, any>; // Todo: Fields should be required
-  fieldsFactory?: (fieldsFactoryParam: any) => Record<string, any>;
+  fields?: QueryFields; // Todo: Fields should be required
+  fieldsFactory?: (fieldsFactoryParam: any) => QueryFields;
 };
